refactor(App): move Legend logic
